+++ src/main/java/Tavi007/ElementalCombat/loading/BiomeCombatProperties.java
@@ -5,15 +5,15 @@ import java.util.HashMap;
 public class BiomeCombatProperties {
 	public static final BiomeCombatProperties EMPTY = new BiomeCombatProperties();
 	
-	private final HashMap<String, Double> defense_element;
+	private final HashMap<String, Integer> defense_element;
 	
 	public BiomeCombatProperties() {
-		this.defense_element = new HashMap<String, Double>();
+		this.defense_element = new HashMap<String, Integer>();
 	}
 
-	public BiomeCombatProperties(HashMap<String, Double> defense_element) {
+	public BiomeCombatProperties(HashMap<String, Integer> defense_element) {
 		this.defense_element = defense_element;
 	}
 	
-	public HashMap<String, Double> getDefenseElement() {return this.defense_element;}
+	public HashMap<String, Integer> getDefenseElement() {return this.defense_element;}
 }
